getSpeedの戻り値変更

+++ include/nnct/interfaces/speed_encoder.hpp
@@ -6,7 +6,7 @@ namespace nnct::interfaces {
 
 class SpeedEncoder {
     public:
-        virtual double getSpeed();
+        virtual int16_t getSpeed();
 };
 
 } // namespace nnct::interfaces
